+++ components/listarEventos/deleteButton.js
@@ -1,31 +1,34 @@
-import React,{useState} from 'react';
-import { Popconfirm } from 'antd';
+import React, { useState, useContext } from 'react';
+import { Popconfirm, message } from 'antd';
 import { DeleteOutlined } from '@ant-design/icons';
 import propTypes from 'prop-types';
-import data  from '../data.json'
+import { cardContext } from '../../context/cardContext';
+import data from '../data.json'; //simulacion de bd
 
-export default function DeleteButton ({ id }) {
+export default function DeleteButton({ id }) {
     const [visible, setVisible] = useState(false);
-    const [confirmLoading, setConfirmLoading] = useState(false);
 
-    const showPopconfirm = () => {
+    const { generarEventos } = useContext(cardContext);
+
+    const showPopconfirm = () => {//
         setVisible(true);
     };
 
     const borrar = () => {
         //peticion a servidor para eleminiar un elemento
-        data.eventos.map((v,i)=> {
-            if(v.id===id)
-            { data.eventos.splice(i,1)}
+        data.eventos.map((v, i) => {
+            if (v.id === id) { data.eventos.splice(i, 1) }
         })
-        setConfirmLoading(false);
     }
 
     const handleOk = () => {
-        setConfirmLoading(true);
+        const config = { key: 'deleteEvent', style: {marginTop: '15vh'}}
+        message.loading({ content: 'Eliminando evento...', ...config });
         setTimeout(() => {
             setVisible(false);
             borrar();
+            generarEventos(); //temporalmente mientras se termina el servidor
+            message.success({ content: 'Evento eliminado', duration: 2, ...config });
         }, 2000);
     };
 
@@ -35,18 +38,16 @@ export default function DeleteButton ({ id }) {
     };
 
     return (
-        <div>
+        <>
             <Popconfirm
-                title="Title"
+                title="Esta acción es irreversible"
                 visible={visible}
                 onConfirm={handleOk}
-                okButtonProps={{ loading: confirmLoading }}
                 onCancel={handleCancel}
             >
-                <DeleteOutlined  onClick={showPopconfirm} />
+                <DeleteOutlined onClick={showPopconfirm} />
             </Popconfirm>
-
-        </div>
+        </>
     )
 }
 DeleteButton.propTypes = {
